Add vitest and update file structure
