+++ src/Bicep.Cli.E2eTests/package.json
@@ -23,6 +23,7 @@
     "typescript": "~5.5.0",
     "typescript-eslint": "^8.8.0",
     "uuid": "^10.0.0",
+    "vitest": "^2.1.2",
     "vscode-jsonrpc": "^8.2.1"
   },
   "scripts": {
