+++ src/Bicep.Cli.E2eTests/src/live/publish.test.ts
@@ -7,10 +7,10 @@
  * @group live
  */
 
-import { BicepRegistryReferenceBuilder } from "./utils/br";
-import { invokingBicepCommand } from "./utils/command";
-import { pathToExampleFile } from "./utils/fs";
-import { getEnvironment } from "./utils/liveTestEnvironments";
+import { BicepRegistryReferenceBuilder } from "../utils/br";
+import { invokingBicepCommand } from "../utils/command";
+import { pathToExampleFile } from "../utils/fs";
+import { getEnvironment } from "../utils/liveTestEnvironments";
 
 describe("bicep publish", () => {
   const testArea = "publish";
