+++ src/Bicep.Cli.E2eTests/src/local/local-deploy.test.ts
@@ -2,50 +2,38 @@
 // Licensed under the MIT License.
 
 /**
- * Live tests for "bicep local-deploy".
+ * Tests for "bicep local-deploy".
  *
- * @group live
+ * @group CI
  */
 
 import os from "os";
-import { invokingBicepCommand } from "./utils/command";
-import { copyToTempFile, pathToExampleFile, pathToTempFile } from "./utils/fs";
+import { invokingBicepCommand } from "../utils/command";
+import { copyToTempFile, pathToExampleFile, pathToTempFile } from "../utils/fs";
 import {
   platformSupportsLocalDeploy,
   publishExtension,
-} from "./utils/localdeploy";
-import { getEnvironment } from "./utils/liveTestEnvironments";
-import { BicepRegistryReferenceBuilder } from "./utils/br";
-import { itif } from "./utils/testHelpers";
+} from "../utils/localdeploy";
+import { itif } from "../utils/testHelpers";
 
 describe("bicep local-deploy", () => {
-  const environment = getEnvironment();
-  const testArea = `local-deploy-live${environment.suffix}`;
-  const builder = new BicepRegistryReferenceBuilder(
-    environment.registryUri,
-    testArea,
-  );
+  const testArea = "local-deploy";
 
   itif(platformSupportsLocalDeploy())(
-    "should publish and run an extension published to a registry",
+    "should publish and run an extension published to the local file system",
     () => {
       const baseFolder = pathToExampleFile("local-deploy");
-      const target = builder.getBicepReference("mock", "0.0.1");
+      const target = pathToTempFile(testArea, "mock.tgz");
 
       const files = {
         bicep: copyToTempFile(baseFolder, "main.bicep", testArea),
         bicepparam: copyToTempFile(baseFolder, "main.bicepparam", testArea),
-        bicepconfig: copyToTempFile(
-          baseFolder,
-          `bicepconfig${environment.suffix}.json`,
-          testArea,
-          {
-            relativePath: "bicepconfig.json",
-            values: {
-              $TARGET_REFERENCE: target,
-            },
+        bicepconfig: copyToTempFile(baseFolder, `bicepconfig.json`, testArea, {
+          relativePath: "bicepconfig.json",
+          values: {
+            $TARGET_REFERENCE: "./mock.tgz",
           },
-        ),
+        }),
       };
 
       const typesIndexPath = pathToTempFile(testArea, "types", "index.json");
