+++ src/Bicep.Cli.E2eTests/vitest.config.ts
@@ -1,0 +1,7 @@
+import { defineConfig } from "vitest/config";
+
+export default defineConfig({
+  test: {
+    testTimeout: 20000,
+  },
+});
